Use copied version of FreeSurfer directory
The FreeSurfer subject directory within $SUBJECTS_DIR is commonly deleted upon partial completion of the protocol, which could make it impossible to run only latter parts of the protocol unmodified. By instead pointing to the version of the FreeSurfer subject directory that has already been duplicated in the output directory, which is likely to be mounted to a host system location, such resumptions are easier.